chore: rename and move web3 zustand store
