+++ libs/deposits/src/lib/deposit-dialog.tsx
@@ -4,7 +4,7 @@ import type { Intent } from '@vegaprotocol/ui-toolkit';
 import { Dialog } from '@vegaprotocol/ui-toolkit';
 import { useCallback, useState } from 'react';
 import { DepositContainer } from './deposit-container';
-import { useWeb3ConnectDialog } from '@vegaprotocol/web3';
+import { useWeb3ConnectStore } from '@vegaprotocol/web3';
 
 interface State {
   isOpen: boolean;
@@ -41,7 +41,7 @@ const DEFAULT_STYLE: DepositDialogStyleProps = {
 
 export const DepositDialog = () => {
   const { assetId, isOpen, open, close } = useDepositDialog();
-  const connectWalletDialogIsOpen = useWeb3ConnectDialog(
+  const connectWalletDialogIsOpen = useWeb3ConnectStore(
     (state) => state.isOpen
   );
   const [dialogStyleProps, _setDialogStyleProps] = useState(DEFAULT_STYLE);
